requirements.txt
streamlit
Pillow
openai
python-dotenv
pandas
datasets
huggingface_hub
